+++ tests/ProjectFilter.UnitTests/Helpers/HierarchyData.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 
 namespace ProjectFilter.Helpers;
@@ -6,10 +8,15 @@ namespace ProjectFilter.Helpers;
 
 internal class HierarchyData {
 
-    public HierarchyData(Guid identifier, string name, Guid? parent) {
+    private Dictionary<string, HierarchyData>? _lookup;
+
+
+    public HierarchyData(Guid identifier, string name, Guid? parent, bool isShared) {
         Identifier = identifier;
         Name = name;
         Parent = parent;
+        IsShared = isShared;
+        DependencyNames = Enumerable.Empty<string>();
     }
 
 
@@ -31,10 +38,31 @@ internal class HierarchyData {
     public bool IsProject { get; private set; }
 
 
+    public bool IsShared { get; }
+
+
+    public IEnumerable<string> DependencyNames { get; private set; }
+
+
+    public IEnumerable<HierarchyData> GetDependencyData() {
+        if (_lookup is not null) {
+            return DependencyNames.Select((x) => _lookup[x]).ToList();
+        }
+
+        return Enumerable.Empty<HierarchyData>();
+    }
+
+
     public void SetType(HierarchyType type, Guid clsid, bool isProject) {
         Type = type;
         CLSID = clsid;
         IsProject = isProject;
+    }
+
+
+    public void SetDependencies(IEnumerable<string> dependencies, Dictionary<string, HierarchyData> lookup) {
+        DependencyNames = dependencies.ToList();
+        _lookup = lookup;
     }
 
 }